# lib/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Fake Media Detection Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DetectionResultPrediction =
  (typeof DetectionResultPrediction)[keyof typeof DetectionResultPrediction];

export const DetectionResultPrediction = {
  Real: "Real",
  Fake: "Fake",
} as const;

export type DetectionResultMediaType =
  (typeof DetectionResultMediaType)[keyof typeof DetectionResultMediaType];

export const DetectionResultMediaType = {
  image: "image",
  video: "video",
  audio: "audio",
} as const;

export interface AnalysisDetails {
  facialInconsistency?: number;
  compressionArtifacts?: number;
  frequencyAnomaly?: number;
  temporalConsistency?: number;
  noisePattern?: number;
  spectralAnomaly?: number;
  voicePitch?: number;
  processingTime: number;
}

export interface DetectionResult {
  id: number;
  prediction: DetectionResultPrediction;
  confidence: number;
  explanation: string;
  mediaType: DetectionResultMediaType;
  filename: string;
  fileSize?: number;
  analysisDetails: AnalysisDetails;
  createdAt: string;
}

export type ScanRecordPrediction =
  (typeof ScanRecordPrediction)[keyof typeof ScanRecordPrediction];

export const ScanRecordPrediction = {
  Real: "Real",
  Fake: "Fake",
} as const;

export type ScanRecordMediaType =
  (typeof ScanRecordMediaType)[keyof typeof ScanRecordMediaType];

export const ScanRecordMediaType = {
  image: "image",
  video: "video",
  audio: "audio",
} as const;

export interface ScanRecord {
  id: number;
  prediction: ScanRecordPrediction;
  confidence: number;
  explanation: string;
  mediaType: ScanRecordMediaType;
  filename: string;
  fileSize?: number;
  analysisDetails?: AnalysisDetails;
  createdAt: string;
}

export interface HistoryResponse {
  scans: ScanRecord[];
  total: number;
  limit: number;
  offset: number;
}

export interface PlatformStats {
  totalScans: number;
  fakeDetected: number;
  realDetected: number;
  imageScans: number;
  videoScans: number;
  audioScans: number;
  accuracyRate: number;
  avgProcessingTime: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type DetectImageBody = {
  file: Blob;
  userId?: string;
};

export type DetectVideoBody = {
  file: Blob;
  userId?: string;
};

export type DetectAudioBody = {
  file: Blob;
  userId?: string;
};

export type GetHistoryParams = {
  userId?: string;
  limit?: number;
  offset?: number;
  mediaType?: GetHistoryMediaType;
};

export type GetHistoryMediaType =
  (typeof GetHistoryMediaType)[keyof typeof GetHistoryMediaType];

export const GetHistoryMediaType = {
  image: "image",
  video: "video",
  audio: "audio",
} as const;