# lib/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Fake Media Detection Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get scan history
      description: Returns list of previous scan results
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: mediaType
          schema:
            type: string
            enum: [image, video, audio]
      responses:
        "200":
          description: Scan history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
  /history/{id}:
    get:
      operationId: getScanById
      tags: [history]
      summary: Get scan by ID
      description: Returns a specific scan result
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AnalysisDetails:
      type: object
      properties:
        facialInconsistency:
          type: number
        compressionArtifacts:
          type: number
        frequencyAnomaly:
          type: number
        temporalConsistency:
          type: number
        noisePattern:
          type: number
        spectralAnomaly:
          type: number
        voicePitch:
          type: number
        processingTime:
          type: number
      required:
        - processingTime
    ScanRecord:
      type: object
      properties:
        id:
          type: integer
        prediction:
          type: string
          enum: [Real, Fake]
        confidence:
          type: number
        explanation:
          type: string
        mediaType:
          type: string
          enum: [image, video, audio]
        filename:
          type: string
        fileSize:
          type: integer
        analysisDetails:
          $ref: "#/components/schemas/AnalysisDetails"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - prediction
        - confidence
        - explanation
        - mediaType
        - filename
        - createdAt
    HistoryResponse:
      type: object
      properties:
        scans:
          type: array
          items:
            $ref: "#/components/schemas/ScanRecord"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - scans
        - total
        - limit
        - offset
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message