# lib/lib/api-zod/src/generated/types/analysisDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Fake Media Detection Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisDetails {
  facialInconsistency?: number;
  compressionArtifacts?: number;
  frequencyAnomaly?: number;
  temporalConsistency?: number;
  noisePattern?: number;
  spectralAnomaly?: number;
  voicePitch?: number;
  processingTime: number;
}