# lib/lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Fake Media Detection Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanRecord } from "./scanRecord";

export interface HistoryResponse {
  scans: ScanRecord[];
  total: number;
  limit: number;
  offset: number;
}