# lib/lib/api-zod/src/generated/types/scanRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Fake Media Detection Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisDetails } from "./analysisDetails";
import type { ScanRecordMediaType } from "./scanRecordMediaType";
import type { ScanRecordPrediction } from "./scanRecordPrediction";

export interface ScanRecord {
  id: number;
  prediction: ScanRecordPrediction;
  confidence: number;
  explanation: string;
  mediaType: ScanRecordMediaType;
  filename: string;
  fileSize?: number;
  analysisDetails?: AnalysisDetails;
  createdAt: Date;
}